requirements.txt
fastapi
uvicorn
pillow
numpy
python-multipart
mysql-connector-python
huggingface_hub
Jinja2
pyttsx3
tensorflow==2.13.0
passlib[bcrypt]
pyjwt
gtts

